splits types for maintainability
