+++ apps/api/src/types/player.ts
@@ -1,0 +1,6 @@
+export type Player = InstanceType<typeof PlayerClass>
+import { Player as PlayerClass } from '../engine/components/Player'
+
+export interface PlayerState {
+  player: Player
+}
